refactor: component name changes + doc comments
